Make application core framework-agnostic (refactoring step)
To get rid of using Spring specific @Autowired annotation where possible.

The idea is to extract the application core logic to a separate domain.
It should be framework-agnostic. There is no need to rely on any framework
for this part. Removing redundant @Autowired annotations should help
in further refactoring and extracting code to separate modules.