Move entities package into the domain
It's hard to find a good solution for extracting the JPA from the core.
So JPA-entity is part of domain for now. By splitting the domain from
entities, there is a risk of loosing lazy-loading opportunities,
clear transaction boundaries, orphan deletion and so on.
For now it's kind of a trade-of by putting JPA within the core-layer.