Extract the application core from infrastructure #16 (#20)
The refactoring step in preparation for domain module extraction.

* Split ports and adapters,
* Cover default method of extracted interface with tests,
* Move repository to the infrastructure package,
* Move entities package into the domain

It's hard to find a good solution for extracting JPA from the core.
So JPA-entity is part of domain for now. By splitting the domain from
entities, there is a risk of loosing lazy-loading opportunities,
clear transaction boundaries, orphan deletion and so on.
For now, it's kind of trade-of by putting JPA within the core-layer.